# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "With this API you can save Todos.",
        "title": "Todo List API",
        "contact": {
            "name": "Alexander Kuepper",
            "email": "kuepper.alexander@web.de"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/todo": {
            "get": {
                "description": "get all Todos",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Show all todos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Todo"
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "description": "you can create a todo with this method",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Create a todo",
                "parameters": [
                    {
                        "description": "todo",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PostTodo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Status Created"
                    },
                    "400": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/todo/{id}": {
            "get": {
                "description": "get todo by ID",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Show a todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Todo"
                        }
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "delete": {
                "description": "everyone can delete the todo",
                "tags": [
                    "todo"
                ],
                "summary": "Delete the todo",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "patch": {
                "description": "with this method you can set the done bit true",
                "tags": [
                    "todo"
                ],
                "summary": "Set todo done",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "model.PostTodo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                }
            }
        },
        "model.Todo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "done": {
                    "type": "boolean"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "priority": {
                    "type": "integer"
                }
            }
        }
    }
}